+++ examples/springboot/spring-boot_app/src/main/java/com/canonical/sampleapp/config/MessagingRabbitmqConfiguration.java
@@ -1,0 +1,26 @@
+package com.canonical.sampleapp.config;
+
+import org.springframework.amqp.core.Queue;
+import org.springframework.amqp.core.TopicExchange;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@ConditionalOnProperty(name = "spring.rabbitmq.host")
+@Configuration
+public class MessagingRabbitmqConfiguration {
+	  static final String topicExchangeName = "";
+	  static final String queueName = "charm";
+
+	  @Bean
+	  Queue queue() {
+	    return new Queue(queueName, false);
+	  }
+
+	  @Bean
+	  TopicExchange exchange() {
+	    return new TopicExchange(topicExchangeName);
+	  }
+
+
+}