+++ examples/springboot/spring-boot_app/src/main/java/com/canonical/sampleapp/service/MessageService.java
@@ -1,0 +1,37 @@
+package com.canonical.sampleapp.service;
+
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.lang.Nullable;
+import org.springframework.stereotype.Component;
+
+import org.springframework.amqp.core.Message;
+import org.springframework.amqp.core.Queue;
+
+@ConditionalOnProperty(name = "spring.rabbitmq.host")
+@Component
+public class MessageService {
+
+    private RabbitTemplate template;
+
+    private Queue queue;
+
+    MessageService(Queue queue, RabbitTemplate template) {
+        this.queue = queue;
+        this.template = template;
+    }
+    
+    public void send(String message) {
+        this.template.convertAndSend(queue.getName(), message);
+        System.out.println(" [x] Sent '" + message + "'");
+    }
+
+    @Nullable
+    public String receive() {
+    	Message message = this.template.receive(queue.getName());
+    	if (message == null) {
+    		return null;
+    	}
+        return new String(message.getBody());
+    }
+}