+++ docs/changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Place any unreleased changes here, that are subject to release in coming versions :).
 
+## 2025-07-15
+
+* feat: Spring Boot support for RabbitMQ.
+
 ## 2025-07-11
 
 * docs: Refactored the RTD tutorial landing page.